moon points up now
Thanks to the NiVZ in the fitbit discord server, the moon is now translated by x and y according to a sinusoidal and cosine curve. This means it always points the right way up, even as a crescent!

there was also some naming issues with the moon pictures. fixed now!